projects/fastAPI-chatbot/requirements.txt
fastapi
uvicorn
pydantic
openai
websocket
python_dotenv
jinja2